Refactor OpenAI client initialization into dedicated module
